[DF-1131] Fixed: getCount now return size_t
Former-commit-id: 71aced0a7ac21b906c738866ade870d2d4718d54

+++ Sources/Internal/Base/EnumMap.cpp
@@ -66,7 +66,7 @@ const char* EnumMap::ToString(const int e) const
 	return ret;
 }
 
-int EnumMap::GetCount() const
+size_t EnumMap::GetCount() const
 {
 	return map.size();
 }
